Feature/member 008 AI 검수

+++ src/main/java/com/earth/ureverse/global/common/dto/response/BrandResponseDto.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class BrandResponseDto {
-    private int brandId;
+    private Long brandId;
     private String name;
     private String nameEn;
     private String logo;
